// management/univention-portal/frontend/src/apis/notifications/api.ts
export interface NotificationRead {
    /**
     *
     * @type {string}
     * @memberof NotificationRead
     */
    'sourceUid': string;
    /**
     *
     * @type {string}
     * @memberof NotificationRead
     */
    'targetUid': string;
    /**
     * The notification title shall be visible to the user in a highlighted way. The title should be kept short.
     * @type {string}
     * @memberof NotificationRead
     */
    'title': string;
    /**
     * The notification details shall be visible to the user below the title. The details can be longer than the title.
     * @type {string}
     * @memberof NotificationRead
     */
    'details': string;
    /**
     * The severity influences how the notification will be displayed to the user. Typically this influences the background color. It does not influence if the notification is shown or not.
     * @type {NotificationSeverity}
     * @memberof NotificationRead
     */
    'severity': NotificationSeverity;
    /**
     * A point in time at which the notification is not relevant anymore. After this point in time the notification shall not be presented to the user anymore.
     * @type {string}
     * @memberof NotificationRead
     */
    'expireTime'?: string;
    /**
     *
     * @type {Link}
     * @memberof NotificationRead
     */
    'link'?: Link;
    /**
     *
     * @type {string}
     * @memberof NotificationRead
     */
    'id': string;
    /**
     *
     * @type {boolean}
     * @memberof NotificationRead
     */
    'popup'?: boolean;
}
// ...
export interface ReceiverApiInterface {
    /**
     * Delete one notification  Allows to delete a notification. This is intended to be used in the following two cases:  1. A *Sender* can delete a notification if it is not valid anymore.    A *Sender* is only allowed to delete notifications which it did create    itself.  2. A *Receiver* can delete a notification if the user does want it to    disappear. A *Receiver* can only delete notifications which are targeted    to the user which it represents.
     * @summary Delete Notification
     * @param {string} id
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ReceiverApiInterface
     */
    deleteNotificationV1NotificationsIdDelete(id: string, options?: AxiosRequestConfig): AxiosPromise<any>;

    /**
     *
     * @summary Get Notification
     * @param {string} id
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ReceiverApiInterface
     */
    getNotificationV1NotificationsIdGet(id: string, options?: AxiosRequestConfig): AxiosPromise<NotificationRead>;

    /**
     * Read the notifications of the current user.
     * @summary Get Notifications
     * @param {string} [limit]
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ReceiverApiInterface
     */
    getNotificationsV1NotificationsGet(limit?: string, options?: AxiosRequestConfig): AxiosPromise<Array<NotificationRead>>;

    /**
     * Flag a notification as hidden.  This will set the attribute `popup` to `false`.
     * @summary Hide Notification
     * @param {string} id
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ReceiverApiInterface
     */
    hideNotificationV1NotificationsIdHidePost(id: string, options?: AxiosRequestConfig): AxiosPromise<any>;

    /**
     *
     * @summary Stream Notifications
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ReceiverApiInterface
     */
    streamNotificationsV1NotificationsStreamGet(options?: AxiosRequestConfig): AxiosPromise<any>;

}

/**
 * ReceiverApi - object-oriented interface
 * @export
 * @class ReceiverApi
 * @extends {BaseAPI}
 */
export class ReceiverApi extends BaseAPI implements ReceiverApiInterface {
    /**
     * Delete one notification  Allows to delete a notification. This is intended to be used in the following two cases:  1. A *Sender* can delete a notification if it is not valid anymore.    A *Sender* is only allowed to delete notifications which it did create    itself.  2. A *Receiver* can delete a notification if the user does want it to    disappear. A *Receiver* can only delete notifications which are targeted    to the user which it represents.
     * @summary Delete Notification
     * @param {string} id
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ReceiverApi
     */
    public deleteNotificationV1NotificationsIdDelete(id: string, options?: AxiosRequestConfig) {
        return ReceiverApiFp(this.configuration).deleteNotificationV1NotificationsIdDelete(id, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     *
     * @summary Get Notification
     * @param {string} id
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ReceiverApi
     */
    public getNotificationV1NotificationsIdGet(id: string, options?: AxiosRequestConfig) {
        return ReceiverApiFp(this.configuration).getNotificationV1NotificationsIdGet(id, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * Read the notifications of the current user.
     * @summary Get Notifications
     * @param {string} [limit]
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ReceiverApi
     */
    public getNotificationsV1NotificationsGet(limit?: string, options?: AxiosRequestConfig) {
        return ReceiverApiFp(this.configuration).getNotificationsV1NotificationsGet(limit, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * Flag a notification as hidden.  This will set the attribute `popup` to `false`.
     * @summary Hide Notification
     * @param {string} id
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ReceiverApi
     */
    public hideNotificationV1NotificationsIdHidePost(id: string, options?: AxiosRequestConfig) {
        return ReceiverApiFp(this.configuration).hideNotificationV1NotificationsIdHidePost(id, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     *
     * @summary Stream Notifications
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ReceiverApi
     */
    public streamNotificationsV1NotificationsStreamGet(options?: AxiosRequestConfig) {
        return ReceiverApiFp(this.configuration).streamNotificationsV1NotificationsStreamGet(options).then((request) => request(this.axios, this.basePath));
    }
}